frontend/app/requirements.txt
flask
pyvis
selenium
python-memcached
backports.csv
prometheus-flask-exporter
